changed custom button to accept more shapes
